1711 - int->long so it can fit pointers in 64-bit
This undoes 1690.

Current constraints on wart's Cells:
  a) each field takes up a multiple of 4 bytes
  b) all Cells are the same size to avoid fragmentation
  c) long and float fit inside pointer

nrefs is an int to save space; no more than 2G incoming pointers.

Mildly weird that intLiterals and toInt deal in longs. It's just the default
integer type.

+++ 004cell.cc
@@ -1,10 +1,12 @@
 //// cell: core lisp data structure with ref-counted garbage collection
 
-unsigned int numAllocs = 0;
+unsigned long numAllocs = 0;
 
 struct Cell {
-  Cell* car;
+  Cell* car; // aliased to long or float
   Cell* cdr;
+
+  // ints save space on 64-bit platforms
   int type;
     #define CONS 0
     #define INTEGER 1
@@ -14,6 +16,7 @@ struct Cell {
     #define TABLE 5
     #define COMPILED_FN 6
   int nrefs;
+
   Cell() :car(NULL), cdr(NULL), type(CONS), nrefs(0) {}
   void init() {
     car=cdr=nil, type=CONS, nrefs=0;
@@ -46,7 +49,7 @@ struct Heap {
 
 Heap* firstHeap = new Heap();
 Heap* currHeap = firstHeap;
-int currCell = 0;
+long currCell = 0;
 Cell* freelist = NULL;
 
 void growHeap() {
@@ -204,9 +207,9 @@ bool equalList(Cell* a, Cell* b) {
       && equalList(cdr(a), cdr(b));
 }
 
-Cell* nthCdr(Cell* x, int n) {
+Cell* nthCdr(Cell* x, long n) {
   Cell* curr = x;
-  for (int idx = n; idx > 0; --idx) {
+  for (long idx = n; idx > 0; --idx) {
     if (!isCons(curr))
       RAISE << "list is too short: " << x << " " << n << endl;
     curr=cdr(curr);
